requirements.txt
google-adk>=1.16.0
google-cloud-aiplatform>=1.124.0
google-genai>=1.40.0
fastapi
pydantic>=2.10.0
python-dotenv>=1.0.0
uvicorn>=0.34.0
google-cloud-bigquery>=3.27.0
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
